Fixed AI move, makeMove, and isValidMove bugs
Fixed bug with AI making a move. Needed to call isValidMove just prior
to makeMove. Fixed this on line 73 of game.cpp

Fixed bug in makeMove that caused tiles to continue flipping even after
an instance of the same color was found

Fixed bug with isValidMove that would incorrectly show a move as valid
if the second space in a given direction was empty, but further on would
have been legal.