Added isValidMove debug check prior to AI making a move
Just prior to AI making its move a final isValidMove check is made. This
is correctly identifying invalid moves.
At higher depth settings the minimax function is returning an invalid
move, but the isValidMove check is not failing. I suspect this has
something to do with how the previous boards are being stored in the
look ahead. It seems like the moves that it tries to make would have
been valid at some point during the look ahead, and it tries to use that
instead of rolling back to the original move that lead down that path.